Heyo, before I tell what I've got I want to apologise for my absence.
I have written up an A* Algorithm in its own class,  the code in the AStar_pathing class should find a path to a target node (this could be attatched to the player, or just the node closest to the player), however I can see some problems when the player is moving that may arise.

Another Issue that I need to address is the contruction of the nodes and how we would prefer to do them, I could write up the nodes in the scene directly and then link them up with the A* class and with the AI, but this would require knowing the specific points in the scene where a node may be required (easy to write up, but need to know exact placements on the scene).

Another solution is to perhaps write up a grid structure which detects where collidable objects are and ignores them for node placement (this would be the more difficult one to write up, but if works correctly should do the job we need).

on a side note I had to comment out the audio manager since there was a dll file missing and I couldn't compile the code.