SAM READ THIS!
Collisions are implemented and working (to some point), however I did comment out a lot of your code and put in another varient you could use.  since i'm not sure how you are going to use the physics for the game I have commented out some the code that was giving errors when I was writing up the new collision code.  now onto the code itself.  Some areas I simiplified quite a bit such as the AABB class.  In the CollisionData class I rewrote the code I had for last year RT3D which does the Collision detection by taking in Collider Objects.  There is a new class CollisionHandler which also has a Colliable class in it which inherits from Collider, this class sets up the AABB and is used to get and set the position of the AABB as well as rotations if we need any.  The Struct in the CollisionHandler is called in Scene to call functions for collision detection.  Got any questions or need help message me.